requirements.txt
anthropic==0.30.1
crem==0.2.12
seaborn==0.13.2
ipywidgets==8.1.3
jupyter==1.0.0
umap-learn==0.5.6
useful-rdkit-utils==0.3.12

# For reinvent4
--extra-index-url https://pypi.anaconda.org/OpenEye/simple
absl-py==1.4.0
alabaster==0.7.13
attrs==22.2.0
babel==2.12.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chemprop==1.5.2
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
docutils==0.19
exceptiongroup==1.1.0
flask==2.2.3
fonttools==4.38.0
funcy==1.18
future==0.18.3
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.1
markdown==3.4.1
markupsafe==2.1.2
matplotlib==3.7.1
mmpdb==2.1
molvs==0.1.1
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
openeye-toolkits==2022.2.2
openeye-toolkits-python3-linux-x64==2022.2.2
packaging==23.0
pandas==1.5.3
pandas-flavor==0.5.0
pathos==0.3.0
pillow==9.4.0
pluggy==1.0.0
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.2
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyyaml==6.0
rdkit==2022.9.5
requests==2.28.2
requests-mock==1.10.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
snowballstemmer==2.2.0
sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
typed-argument-parser==1.7.2
typing-extensions==4.7
typing-inspect==0.8.0
urllib3==1.26.14
werkzeug==2.2.3
wheel==0.38.4
xarray==2023.2.0
xxhash==3.4.1
